requirements.txt
# Core dependencies
chromadb>=0.4.0
openai>=1.0.0
python-dotenv>=1.0.0
numpy>=1.24.0

# Data processing
datasets>=2.14.0  # For HuggingFace datasets
tiktoken>=0.5.0  # For token counting
pandas>=2.0.0

# LLM orchestration
langgraph>=0.0.20
langchain>=0.1.0
langchain-openai>=0.0.5

# Async support
aiohttp>=3.9.0
asyncio>=3.4.3

# Utilities
tqdm>=4.66.0
pydantic>=2.0.0
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development
ipython>=8.0.0
jupyter>=1.0.0
black>=23.0.0
flake8>=6.0.0